# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "WEBMASTERY SCHOOL API DOCS",
        "description": "Webmastery School Portal API documentation",
        "contact": {
            "email": "narushevich.maksim@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:7000/api/v1",
            "description": "API documentation"
        }
    ],
    "paths": {
...
        "/lessons": {
            "get": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Get all lessons",
                "description": "Returns all lessons details",
                "operationId": "getLessons",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Lesson page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items to be retrieved per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "relationships",
                        "in": "query",
                        "description": "Relationships that should be inclused in response",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "videos;tasks;comments;user"
                    },
                    {
                        "name": "relationPerPage",
                        "in": "query",
                        "description": "Number of relationships to be retrieved in response structure",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully received available lessons",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Create a new lesson",
                "description": "Create a new lesson.",
                "operationId": "createLesson",
                "requestBody": {
                    "description": "Create lesson",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateLesson"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lesson created"
                    },
                    "400": {
                        "description": "Request validation error"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lessons/{id}": {
            "get": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Get a specific lesson",
                "description": "Returns lesson details",
                "operationId": "getLessonById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully received specific lessons",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lesson"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    },
                    "404": {
                        "description": "Lesson Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Delete lesson",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteLesson",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of lesson to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lesson deleted"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    },
                    "404": {
                        "description": "Lesson not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Update a specific lesson",
                "description": "Update lesson.",
                "operationId": "updateLesson",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update lesson",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Lesson"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson updated"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    },
                    "400": {
                        "description": "Request validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lessons/{id}/tags": {
            "put": {
                "tags": [
                    "Lesson"
                ],
                "summary": "Update lesson's tags",
                "description": "Update lesson's tags.",
                "operationId": "updateLessonTags",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Lesson id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Lesson's tags",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tags"
                                ],
                                "properties": {
                                    "tags": {
                                        "description": "Lesson's tags",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "action": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "string",
                                        "example": "ATTACH"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lesson tags updated"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    },
                    "400": {
                        "description": "Request validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/videos": {
            "get": {
                "tags": [
                    "Video"
                ],
                "summary": "Get all videos",
                "description": "Returns all videos details",
                "operationId": "getVideos",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Video page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of items to be retrieved per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 10
                    },
                    {
                        "name": "relationships",
                        "in": "query",
                        "description": "Relationships that should be inclused in response",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "lesson;user"
                    },
                    {
                        "name": "relationPerPage",
                        "in": "query",
                        "description": "Number of relationships to be retrieved in response structure",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 3
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully received available videos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Video"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Video"
                ],
                "summary": "Create a new video",
                "description": "Create a new video.",
                "operationId": "createLesson",
                "requestBody": {
                    "description": "Create video",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateVideo"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Video created"
                    },
                    "400": {
                        "description": "Request validation error"
                    },
                    "401": {
                        "description": "Authorization token must be provided"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Lesson": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "course_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CreateLesson": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Lesson"
                    },
                    {
                        "required": [
                            "name",
                            "description",
                            "course_id"
                        ]
                    }
                ]
            },
...
            "Video": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "link": {
                        "type": "string"
                    },
                    "lesson_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "CreateVideo": {
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/Video"
                    },
                    {
                        "required": [
                            "title",
                            "description",
                            "link",
                            "lesson_id"
                        ]
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Oauth2 security",
                "name": "oauth2",
                "in": "header",
                "flows": {
                    "clientCredentials": {
                        "tokenUrl": "oauth/token",
                        "scopes": {}
                    }
                },
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}